+++ Shared/DTOs/PostAsset.cs
@@ -8,7 +8,7 @@ namespace Shared.DTOs
 {
     public  class PostAsset
     {
-        public int BarCode { get; set; }
+        public long BarCode { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
         public string? Note { get; set; }
@@ -18,7 +18,7 @@ namespace Shared.DTOs
     public class GetAsset
     {
         public int Id { get; set; }
-        public int BarCode { get; set; }
+        public long  BarCode { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
         public string? Note { get; set; }
